refactor: update all typo
